pyproject.toml
[tool.poetry]
name = "repl-nix-deadbeesociety-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
flask-login = "^0.6.3"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
flask-socketio = "^5.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
